Update Universal Dependencies, move utils to _utils and fix ngc build
