+++ schematics/universal/package.json
@@ -2,7 +2,7 @@
   "author": "Maciej Treder <contact@maciejtreder.com>",
   "name": "@ng-toolkit/universal",
   "main": "dist/index.js",
-  "version": "7.1.2",
+  "version": "8.0.0",
   "description": "Adds Angular Universal support for any Angular CLI project",
   "repository": {
     "type": "git",
@@ -31,36 +31,34 @@
   "license": "MIT",
   "schematics": "./collection.json",
   "peerDependencies": {
-    "@angular/common": "^7.1.0",
-    "@angular/core": "^7.1.0"
+    "@angular/common": "^8.2.6",
+    "@angular/core": "^8.2.6"
   },
   "dependencies": {
-    "@ng-toolkit/_utils": "7.1.2",
-    "@nguniversal/express-engine": "^7.1.1",
-    "bugsnag": "^2.4.2"
+    "@angular-devkit/core": "^8.3.3",
+    "@angular-devkit/schematics": "^8.3.3",
+    "@ng-toolkit/_utils": "8.0.0",
+    "@nguniversal/express-engine": "^8.1.1",
+    "@schematics/angular": "^8.3.3",
+    "bugsnag": "^2.4.3"
   },
   "devDependencies": {
-    "@ng-toolkit/pwa": "^7.1.1",
-    "@ng-toolkit/serverless": "^7.1.2",
-    "@angular-devkit/build-angular": "^0.13.6",
-    "@angular-devkit/build-ng-packagr": "^0.13.6",
-    "@angular-devkit/core": "^7.1.0",
-    "@angular-devkit/schematics": "^7.1.0",
-    "@angular/cli": "^7.1.0",
-    "@angular/common": "^7.1.0",
-    "@angular/compiler": "^7.1.0",
-    "@angular/compiler-cli": "^7.1.0",
-    "@angular/core": "^7.1.0",
-    "@schematics/angular": "^7.1.0",
-    "@types/jasmine": "^2.6.0",
-    "@types/node": "^8.10.17",
+    "@angular-devkit/build-angular": "^0.803.4",
+    "@angular-devkit/build-ng-packagr": "^0.803.4",
+    "@angular/cli": "^8.3.3",
+    "@angular/common": "^8.2.6",
+    "@angular/compiler": "^8.2.6",
+    "@angular/compiler-cli": "^8.2.6",
+    "@angular/core": "^8.2.6",
+    "@types/jasmine": "^3.4.0",
+    "@types/node": "^12.7.4",
     "ci-publish": "^1.3.1",
-    "cp-cli": "^1.1.2",
-    "jasmine": "^2.8.0",
-    "ng-packagr": "^3.0.3",
-    "rxjs": "^6.3.3",
-    "tsickle": "^0.33.1",
-    "typescript": "~3.2.4"
+    "cp-cli": "^2.0.0",
+    "jasmine": "^3.4.0",
+    "ng-packagr": "^5.5.0",
+    "rxjs": "^6.5.3",
+    "tsickle": "^0.37.0",
+    "typescript": "~3.5.3"
   },
   "publishConfig": {
     "access": "public"
