+++ schematics/universal/tsconfig.json
@@ -1,37 +1,11 @@
 {
-  "compilerOptions": {
-    "baseUrl": "tsconfig",
-    "lib": [
-      "es2017",
-      "dom"
-    ],
-    "declaration": true,
-    "module": "commonjs",
-    "moduleResolution": "node",
-    "noEmitOnError": true,
-    "noFallthroughCasesInSwitch": true,
-    "noImplicitAny": true,
-    "noImplicitThis": true,
-    "noUnusedParameters": true,
-    "noUnusedLocals": true,
-    "declarationDir": "src/",
-    "outDir": "dist/",
-    "skipDefaultLibCheck": true,
-    "skipLibCheck": true,
-    "sourceMap": true,
-    "strictNullChecks": true,
-    "target": "es6",
-    "types": [
-      "jasmine",
-      "node"
-    ]
-  },
+  "extends": "../tsconfig.json",
+  "include": [
+    "./**/*"
+  ],
   "exclude": [
     "./**/files/**/*",
     "./**/ngt-universal/**",
     "./**/public_api.ts"
-  ],
-  "include": [
-    "./**/*"
   ]
 }
