+++ schematics/universal/tsconfig.lib.json
@@ -1,5 +1,6 @@
 {
   "compilerOptions": {
+    "baseUrl": "./",
     "outDir": "./out-tsc/lib",
     "target": "es2015",
     "module": "es2015",
